Added ability to exclude GUI client library requirements from cmake configuration. Fixed an invalid call to min() casting unsigned 8-bit to plain ol' int.
git-svn-id: svn://svn.code.sf.net/p/segs/code2@133 a19be42f-918d-4297-8077-ab25d1e4f51f

+++ Projects/CoX/Servers/MapServer/src/Entity.cpp
@@ -274,7 +274,7 @@ MobEntity::MobEntity()
 }
 void Entity::sendXLuency(BitStream &bs,float val) const
 {
-	storeBitsConditional(bs,8,min(u8(val*255),255)); // upto here everything is ok
+	storeBitsConditional(bs,8,min(static_cast<int>(u8(val*255)),255)); // upto here everything is ok
 }
 void Entity::sendTitles(BitStream &bs) const
 {
